apps/ml-daemon/requirements.txt
anyio==3.6.2
certifi==2022.12.7
dnspython==2.3.0
GPUtil==1.4.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Pillow==9.4.0
psutil==5.9.4
py-cpuinfo==9.0.0
pystray==0.19.4
python-dotenv==1.0.0
PyYAML==6.0
rfc3986==1.5.0
schedule==1.1.0
pyyaml==6.0
six==1.16.0
sniffio==1.3.0
